Add category menu
I want to add a selection menu for categories when the user
creates a new idea.

I don't have a test for this yet, but will when I find out how
to test with selection menus in rspec.

I also want an admin to be able to delete a category.
This is built and there is a test for it.